requirements.txt
bioblend==0.9.0
boto3==1.4.4
celery==3.1.26.post2
certifi==2018.11.29
cffi==1.9.1
codecov==2.0.11
coverage==4.0.3
cryptography==2.5
Django==1.8.18
django-celery==3.2.2
django-chunked-upload==1.1.1
django-cuser==2017.3.16
django-debug-toolbar==1.9.1
django-docker-engine==0.1.4
django-extensions==1.6.7
django-flatblocks==0.9.4
django-guardian==1.4.9
django-haystack==2.7.0
django-markwhat==1.6.1
django-override-storage==0.1.6
django-registration-redux==1.11
djangorestframework==3.3.3
djangorestframework-recursive==0.1.2
django-storages==1.6.0
factory_boy==2.8.1
flake8==2.5.4
flower==0.9.1
jsonschema==2.6.0
Markdown==2.2.1
mock==2.0.0
git+git://github.com/lunant/mockcache@master
networkx==1.7
psycopg2-binary==2.7.4
pycparser==2.13
Pygments==1.6rc1
pyOpenSSL==17.5.0
pysam==0.15.3
pysolr==3.3.2
python-magic==0.4.6
python-memcached==1.59
readline==6.2.2
requests==2.21.0
six==1.10.0
supervisor==4.0.0
unidecode==0.4.21
uuid==1.30
xmltodict==0.9.2
